Move color_chooser, style_chooser, and filter_popup
As plot_api is used by both ert.gui.plottery and ert.gui.tools.plot, its
current location causes circular dependency problems. This PR moves
plottery into ert.gui.tools.plot so that the only exposed symbol from
ert.gui.tools.plot is PlotTool.
