+++ src/ert/gui/tools/plot/__init__.py
@@ -1,13 +1,5 @@
-from .color_chooser import ColorBox
-from .filter_popup import FilterPopup
-from .filterable_kw_list_model import FilterableKwListModel
 from .plot_tool import PlotTool
-from .style_chooser import StyleChooser
 
 __all__ = [
-    "ColorBox",
-    "FilterPopup",
-    "FilterableKwListModel",
     "PlotTool",
-    "StyleChooser",
 ]
