+++ src/ert/gui/tools/plot/customize/color_chooser.py
