+++ src/ert/gui/tools/plot/customize/style_customization_view.py
@@ -2,10 +2,9 @@ from typing import TYPE_CHECKING, List, Tuple
 
 from qtpy.QtWidgets import QHBoxLayout
 
-from ert.gui.tools.plot import ColorBox
-from ert.gui.tools.plot import style_chooser as sc
-
+from .color_chooser import ColorBox
 from .customization_view import CustomizationView, WidgetProperty
+from .style_chooser import STYLESET_TOGGLE
 
 if TYPE_CHECKING:
     from ert.gui.tools.plot.plottery import PlotConfig
@@ -34,7 +33,7 @@ class StyleCustomizationView(CustomizationView):
             "observs_style",
             "Observation",
             "Line and marker style for the observation line.",
-            line_style_set=sc.STYLESET_TOGGLE,
+            line_style_set=STYLESET_TOGGLE,
         )
 
         self["default_style"].createLabelLayout(layout)
