+++ src/ert/gui/tools/plot/widgets/__init__.py
@@ -1,5 +1,6 @@
 from .clearable_line_edit import ClearableLineEdit
 from .copy_style_to_dialog import CopyStyleToDialog
 from .custom_date_edit import CustomDateEdit
+from .filter_popup import FilterPopup
 
-__all__ = ["ClearableLineEdit", "CopyStyleToDialog", "CustomDateEdit"]
+__all__ = ["ClearableLineEdit", "CopyStyleToDialog", "CustomDateEdit", "FilterPopup"]
